requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
cssselect2==0.8.0
Django==5.1.7
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
idna==3.10
Markdown==3.8
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-barcode==0.15.1
python-dotenv==1.0.1
qrcode==8.2
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.4.0
webencodings==0.5.1
zopfli==0.2.3.post1
